+++ example/lib/main.dart
@@ -96,7 +96,11 @@ class _MyHomePageState extends State<MyHomePage> {
                   'Animations : $availableAnimations --- Length : ${availableAnimations.length}');
               chosenAnimation = await showPickerDialog(
                   'Animations', availableAnimations, chosenAnimation);
-              controller.playAnimation(animationName: chosenAnimation);
+              //Play animation with loop count
+              controller.playAnimation(
+                animationName: chosenAnimation,
+                loopCount: 2,
+              );
             },
             icon: const Icon(Icons.format_list_bulleted_outlined),
           ),