+++ lib/src/controllers/i_flutter_3d_controller.dart
@@ -2,7 +2,14 @@ abstract class IFlutter3DController {
   /// Causes animations to be played, Can be use to switch animations as well.
   /// If animationName is null and model has at list one animation, it will play first model's animation
   /// If animationName passed and not null it will play specific animation.
-  void playAnimation({String? animationName});
+  ///
+  /// If loopCount > 0, the animation will loop for the specified count.
+  /// If loopCount == 0, the animation will loop infinitely.
+  /// (Default is 0)
+  void playAnimation({
+    String? animationName,
+    int loopCount = 0,
+  });
 
   ///Causes animations to be paused at current frame.
   void pauseAnimation();